requirements.txt
aniso8601==9.0.1
click==8.0.3
Flask==2.0.2
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
myWallet==0.2
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
